requirements.txt
falcon
inflect
audioread
Unidecode
librosa
matplotlib
numpy
scipy
tqdm
keras
msgpack
msgpack_numpy
tensorflow_gpu